refactor: make server and client using same function to get user move function
