+++ Javascript/Fruits/game.js
@@ -1,3 +1,4 @@
+const  playerMover = require('./public/move-player-rule')
 
 exports.createGame = function createGame() {
 
@@ -94,44 +95,24 @@ exports.createGame = function createGame() {
 
     function movePlayer(command) {
 
-        const acceptedCommands = {
-            ArrowUp(player) {
-                if (player.y - 1 >= 0) {
-                    player.y = player.y - 1
-                }
-            },
-            ArrowDown(player) {
-
-                if (player.y + 1 < state.screen.height) {
-                    player.y = player.y + 1
-                }
-            },
-            ArrowLeft(player) {
-                if (player.x - 1 >= 0) {
-                    player.x = player.x - 1
-                }
-            },
-            ArrowRight(player) {
-                if (player.x + 1 < state.screen.width) {
-                    player.x = player.x + 1
-                }
-            }
-        }
-
         console.log(`Moving ${command.playerId} with ${command.keyPressed}`)
 
         const { keyPressed, playerId } = command
 
         const player = state.players[playerId]
 
-        const action = acceptedCommands[keyPressed]
+        const action = playerMover.movePlayerRule(keyPressed, state.screen)
 
         if (player && action) {
-            action(player)
+            const playerOnNewPosition = action(player)
+
+            this.state.players[playerOnNewPosition.id] = {
+                ...playerOnNewPosition
+            }
 
             notifyAll({
                 type: 'move-player',
-                ...player
+                ...playerOnNewPosition
             })
 
             checkForFruitCollision(playerId)
